The program now works!
To exceed the requirements, this program:
Loads scriptures from a .txt file
Allows users to choose a scripture or be given a random one
Users can choose from hiding all letters of words that are hidden, or leave the first letter unhidden.
Punctuation is treated as words that are never hidden, and the program ends when all of the regular words are hidden.
The user is greeted when the program begins
The user is congruadulated when they complete the verse by hiding all of the words and then pressing 'enter' once again.